+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 - Add `Venue.fullAddress` (#3399)
+- Include the webinar URL in the calendar invite (#3445)
 - Add a field for the webinar URL to events (#3380, #3384)
 - Add more venue fields: contact person, email, phone number (#3362, #3363)
 - Add event formats (on-site, hybrid, online) (#3343, #3361)